requirements.txt
Flask==2.3.2
matplotlib
numpy
Pillow
fuzzywuzzy
python-Levenshtein